Performance improvements to get_monthly_user_activity()
We have a 33K users Moodle install. and for us to be able to run this (beautiful) report on a per course basis, without timeout errors, we needed to narrow the graphs population to the users of that specific course. (It's seems also the right thing to do, regardless performance)
So I am adding the suggested patch to include only the users of the specific course being viewed.
Please review and see if you can integrate.
Kindly, 
Nadav